requirements.txt
anthropic==0.49.0
httpx==0.28.1
python-dotenv==1.1.0
supabase==2.15.1
reportlab==4.4.0
Pillow==12.2.0
